% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alarms.R
\name{delete_alarms}
\alias{delete_alarms}
\alias{get_alarms}
\alias{get_alarm_history}
\alias{temp_alarm_state}
\title{CloudWatch Alarms}
\usage{
delete_alarms(alarm, ...)

get_alarms(alarm, prefix, state, n, token, ...)

get_alarm_history(alarm, type, start, end, n, token, ...)

temp_alarm_state(alarm, reason, state, ...)
}
\arguments{
\item{alarm}{A character vector specifying one or more CloudWatch alarm names.}

\item{\dots}{Additional arguments passed to \code{\link{cloudwatchHTTP}}.}

\item{prefix}{Rather than specify \code{alarm}, it is possible to specify the \emph{prefix} of a set of alarm names using this parameter.}

\item{state}{An optional character string specifying an alarm state to be used when matching alarms. One of: \dQuote{OK}, \dQuote{ALARM}, or \dQuote{INSUFFICIENT_DATA}.}

\item{n}{The maximum number of records to return.}

\item{token}{A pagination token.}

\item{type}{Optionally, a character string specifying the type of alarm history to retrieve. One of: \dQuote{ConfigurationUpdate}, \dQuote{StateUpdate}, or \dQuote{Action}.}

\item{start}{Optionally, the start date of alarm history to retrieve.}

\item{end}{Optionally, the end date of alarm history to retrieve.}

\item{reason}{A charater string specifying the alarm is being temporarily set to this state.}
}
\description{
Get CloudWatch alarms and alarm information; delete alarms
}
\details{
These functions delete, modify, and retrieve details about CloudWatch alarms. \code{create_alarms} creates a new alarm. \code{delete_arms} deletes one or more alarms. \code{get_alarms} can retrieve a specific alarm or list available alarms, and \code{get_alarm_history} retrieve's an alarms history. \code{temp_alarm_state} temporarily changes an alarm's status, which can be useful for testing purposes.
}
\seealso{
\code{\link{enable_actions}}
}
