\name{aws.cloudwatch-package}
\alias{aws.cloudwatch-package}
\alias{aws.cloudwatch}
\docType{package}
\title{AWS CloudWatch Client Package}
\description{A simple client package for the Amazon Web Services AWS CloudWatch REST API.}
\details{}
\author{Thomas J. Leeper <thosjleeper@gmail.com>}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}
