% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{cloudwatchHTTP}
\alias{cloudwatchHTTP}
\title{cloudwatchHTTP}
\usage{
cloudwatchHTTP(query, api = c("monitoring", "events", "logs"),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), ...)
}
\arguments{
\item{query}{A list.}

\item{api}{A character string specifying which CloudWatch API to call. One of \dQuote{monitoring}, \dQuote{events}, or \dQuote{logs}.}

\item{region}{A character string containing the AWS region. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. If missing, defaults to value stored in environment variable \dQuote{AWS_ACCESS_KEY_ID}.}

\item{secret}{A character string containing an AWS Secret Access Key. If missing, defaults to value stored in environment variable \dQuote{AWS_SECRET_ACCESS_KEY}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\description{
Execute CloudWatch HTTP Requests
}
\details{
This is the workhorse function to execute CloudWatch API requests. It should not normally be used directly.
}

