% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_events.R
\name{create_log_events}
\alias{create_log_events}
\alias{get_log_events}
\title{Log events}
\usage{
create_log_events(name, stream, events, token, ...)

get_log_events(name, stream, start, end, ascending, n, token, ...)
}
\arguments{
\item{name}{A character string specifying a log group name.}

\item{stream}{A character string specifying a log stream name.}

\item{events}{This is a list of two-element lists, where each list contains a character string named \code{message} and a unix epoch time in milliseconds (e.g., \code{list(message = "hello world!", timestamp = 1)}).}

\item{token}{A pagination token.}

\item{start}{Unix epoch time in milliseconds.}

\item{end}{Unix epoch time in milliseconds.}

\item{ascending}{A logical specifying whether earliest events should be returned first. Default is implicitly \code{FALSE}.}

\item{n}{The maximum number of records to return.}

\item{\dots}{Additional arguments passed to \code{\link{cloudwatchHTTP}}.}
}
\description{
Get and create log events
}
\details{
\code{get_log_events} retrieves event data. \code{create_log_events} adds event data to the log.
}

