% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{create_log_group}
\alias{create_log_group}
\alias{delete_log_group}
\alias{get_log_groups}
\title{Cloudwatch log groups}
\usage{
create_log_group(name, ...)

delete_log_group(name, ...)

get_log_groups(prefix, n, token, ...)
}
\arguments{
\item{name}{A character string specifying a log group name.}

\item{prefix}{Optionally, a log stream \emph{prefix} used to restrict the returned results.}

\item{n}{The maximum number of records to return.}

\item{token}{A pagination token.}

\item{\dots}{Additional arguments passed to \code{\link{cloudwatchHTTP}}.}
}
\description{
Create, retrieve, and delete log groups
}

