% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_streams.R
\name{create_log_stream}
\alias{create_log_stream}
\alias{delete_log_stream}
\alias{get_log_streams}
\title{Log streams}
\usage{
create_log_stream(name, stream, ...)

delete_log_stream(name, stream, ...)

get_log_streams(name, prefix, ascending, sort, n, token, ...)
}
\arguments{
\item{name}{A character string specifying a log group name.}

\item{stream}{A character string specifying a log stream name.}

\item{prefix}{Optionally, a log stream \emph{prefix} used to restrict the returned results.}

\item{ascending}{A logical specifying whether earliest events should be returned first. Default is implicitly \code{TRUE}.}

\item{sort}{Optionally, a character string specifying the sort order of the results. Must be one of: \dQuote{LogStreamName} or \dQuote{LastEventTime}.}

\item{n}{The maximum number of records to return.}

\item{token}{A pagination token.}

\item{\dots}{Additional arguments passed to \code{\link{cloudwatchHTTP}}.}
}
\description{
Get, create, and delete log streams
}

