% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_filters.R
\name{create_metric_filter}
\alias{create_metric_filter}
\alias{delete_metric_filter}
\alias{get_metric_filters}
\alias{test_metric_filter}
\title{Metric Filters}
\usage{
create_metric_filter()

delete_metric_filter(name, filter, ...)

get_metric_filters(name, prefix, n, token, ...)

test_metric_filter(pattern, messages, ...)
}
\arguments{
\item{name}{A character string specifying a log group name.}

\item{filter}{The name of a metric filter.}

\item{prefix}{A filter name \emph{prefix} used to restrict the returned results.}

\item{n}{The maximum number of records to return.}

\item{token}{A pagination token.}

\item{pattern}{According to the API documentation: \dQuote{A symbolic description of how CloudWatch Logs should interpret the data in each log event. For example, a log event may contain timestamps, IP addresses, strings, and so on. You use the filter pattern to specify what to look for in the log event message.}}

\item{messages}{A character vector of event messages.}

\item{\dots}{Additional arguments passed to \code{\link{cloudwatchHTTP}}.}
}
\description{
Get, create, delete, and test metric filters
}

