% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention_policies.R
\name{create_retention}
\alias{create_retention}
\alias{delete_retention}
\title{Retention policies}
\usage{
create_retention(name, days, ...)

delete_retention(name, ...)
}
\arguments{
\item{name}{A character string specifying a log group name.}

\item{days}{An integer specifying a number of days to retain logs for. There are a finite allowed set of values (see API documentation).}

\item{\dots}{Additional arguments passed to \code{\link{cloudwatchHTTP}}.}
}
\description{
Create and delete retention policies
}

