% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.cloudwatch-package.R
\docType{package}
\name{aws.cloudwatch-package}
\alias{aws.cloudwatch-package}
\alias{aws.cloudwatch}
\title{aws.cloudwatch}
\description{
AWS CloudWatch Client Package
}
\details{
A simple client package for the Amazon Web Services AWS CloudWatch REST APIs. The package wraps three separate APIs: the CloudWatch API (which is used to manage CloudWatch alarms), Events API, and Logs API. CloudWatch is useful for monitoring use of AWS resources (such as creating a billing alarm and corresponding automatic reactions) or setting up logging and/or notifications in response to particular events on AWS resources.
}
\references{
\url{http://docs.aws.amazon.com/AmazonCloudWatch/latest/DeveloperGuide/WhatIsCloudWatch.html}, \url{http://docs.aws.amazon.com/AmazonCloudWatchEvents/latest/APIReference/Welcome.html}, \url{http://docs.aws.amazon.com/AmazonCloudWatch/latest/DeveloperGuide/WhatIsCloudWatchLogs.html}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
