% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.cloudwatch-package.R
\docType{package}
\name{aws.cloudwatch-package}
\alias{aws.cloudwatch}
\alias{aws.cloudwatch-package}
\title{aws.cloudwatch}
\description{
AWS CloudWatch Client Package
}
\details{
A simple client package for the Amazon Web Services AWS CloudWatch REST API.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}

